Conflicts:
	Simulation/docs/doc.pdf

+++ Simulation/src/ste/framework_graphics/shadows/screen_space_soft_shadows/shaders/ssss_blur_y.frag
@@ -1,10 +1,11 @@
 
 #type frag
 #version 450
+#extension GL_NV_gpu_shader5 : require
 
 #include "ssss_blur.glsl"
 
-out mediump float gl_FragColor;
+out float gl_FragColor;
 
 layout(binding = 7) uniform sampler2D z_buffer;
 layout(binding = 8) uniform sampler2DArray src;
